Criação da biblioteca lista
