+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
   "files.associations": {
-    "fisica_calc.h": "c"
+    "fisica_calc.h": "c",
+    "eletrotastica_calc.h": "c"
   },
   "C_Cpp_Runner.cCompilerPath": "gcc",
   "C_Cpp_Runner.cppCompilerPath": "g++",
